Tibia: Fix malformed packet exceptions on first pass
The first pass and other passes with NULL proto_tree caused a malformed packet
exception. This was due to an outdated if (tree) check that is now removed.

While we are at it, adjust documentation to reflect the actual maximum version
supported: 11.00 and add the new (yet unused) version traits.

Bug: 13959
Fixes: g62c9a8a865
Change-Id: If3e1093c881c0eb5070fefc9877cf5f0c61a7572
Reviewed-on: https://code.wireshark.org/review/23446
Petri-Dish: Michael Mann <mmann78@netscape.net>
Reviewed-by: Michael Mann <mmann78@netscape.net>