prefs: Fix leak when parsing empty pref string
Fix minor leak. When parsing preference string lists and preference
is empty a buffer is allocated to hold the string but is then never
inserted into the string list as it is empty. This causes a minor leak
as no reference is kept to allocated buffer and won't be freed by
corresponding clear string list function call.

Bug: 14071
Change-Id: I1edcc77095c0f430e03a49491e5281730fbceb95
Reviewed-on: https://code.wireshark.org/review/23598
Reviewed-by: Michael Mann <mmann78@netscape.net>
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>