stats_tree: fix the comparison routine for COL_AVERAGE
Generally, the average is calculated as node->total / node->count.

The curent code does not handle the case where we compare two nodes
and both have count == 0. It defines one of the nodes to be bigger.
This triggers (at least on Windows) an assertion about invalid operator<.

To fix this, we define average = 0 for a node with count == 0.
We can then simply compare the two averages.

Change-Id: Ie7d9cd590deddcdb9214c4a2693c2eb47c66b287
Reviewed-on: https://code.wireshark.org/review/23799
Reviewed-by: Martin Kaiser <wireshark@kaiser.cx>
Petri-Dish: Martin Kaiser <wireshark@kaiser.cx>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>