Remove the SVR4 packaging assets.
Remove the svr4-package and solaris-package targets along with their
associated files and directories. We used to use this to build Solaris
packages but we haven't shipped those in years. Given that the last
substantive change to packaging/svr4 was in 2008 it's likely that this
has been unused for a while.

Change-Id: Ib9153c99f503200ea8c48d3ef81ad688ee55c09f
Reviewed-on: https://code.wireshark.org/review/23808
Petri-Dish: Gerald Combs <gerald@wireshark.org>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>