Update comment.
Red Hat, at one point, provided a /usr/include/pcap/pcap.h without
bothering to also provide a /usr/include/pcap.h that linked to it or
included it, breaking source compatibility.  That's what we're working
around when we search in both the top-level include directories and, if
they exist, pcap subdirectories of those directories.

libpcap 1.0 and later also put pcap.h in a pcap subdirectory, but also
provided a pcap.h in the top-level include directory that included
pcap/pcap.h, preserving source compatibility, so that's not the reason
we're searching in the top-level include directories and their pcap
subdirectories.

Change-Id: I8f427d46ce8293d278be9005ee623cda1ea5d691
Reviewed-on: https://code.wireshark.org/review/23902
Reviewed-by: Guy Harris <guy@alum.mit.edu>