autotools: Build and install user guide
Installs the WSUG to $docdir, unless the --disable-guides option is given
to the configure script.

This will open the Help->Contents menu action using the locally installed
guide. It will also point to the correct version for this Wireshark release.
The wireshark.org link only provides the guide for the master branch.

Change-Id: I9dd539c737cb3777dc1351707b9e45a3a6764aaa
Reviewed-on: https://code.wireshark.org/review/23906
Petri-Dish: João Valverde <j@v6e.pt>
Reviewed-by: João Valverde <j@v6e.pt>
Tested-by: João Valverde <j@v6e.pt>
Reviewed-by: Michael Mann <mmann78@netscape.net>