dissector/mqtt: Add the MQTT protocol description
The packet-PROTOABBREV.c template recommends to provide a short
description of the protocol below the license header. Currently,
this information is not present in the packet-mqtt.c dissector.

This patch adds the protocol description taken from the official
specification. Links to the v3.1 and v3.1.1 specifications are
also provided by this patch.

Change-Id: I9bb85aa3b78c8804c923f77c163904a7949f6899
Signed-off-by: Flavio Santes <flavio.santes@1byt3.com>
Reviewed-on: https://code.wireshark.org/review/23936
Reviewed-by: Stig Bjørlykke <stig@bjorlykke.org>
(cherry picked from commit 5967621d19204947021932ea86f1e55fbe810124)
Reviewed-on: https://code.wireshark.org/review/24027