Use "not installed for development" for libgcrypt.
This is similar phrasing to what I used for some other libraries;
hopefully it will keep people from asking "Why am I getting this error?
I installed libgcrypt" questions by making it clear that "installing
libgcrypt" might not be enough.

Also, don't give a package name, because the package might not be called
"libgcrypt-devel" - it might be called "libgcrypt20-dev" of something
such as that.

Change-Id: I486a239e346955666d08cad3b1f8e3a961120e76
Reviewed-on: https://code.wireshark.org/review/24052
Reviewed-by: Guy Harris <guy@alum.mit.edu>