Autodetect more AMQP 0-9 captures
When an capture file containing AMQP 0-9 traffic is missing the
protocol header, and the first packet contains more than one AMQP
frame, we'd fail to detect the protocol version properly, since we'd
check for the frame size being equal to the PDU size.  Fix this case
by checking if the frame fits within the PDU instead.

While it's possible that 0-10 and 1.0 captures could be misdetected as
0-9, this seems highly unlikely.  See details in comment in
packet-amqp.c.

Bug: 14113
Change-Id: I98978ed3919be3daf7fe9a972b87e09f27ba20e3
Reviewed-on: https://code.wireshark.org/review/24108
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot
Reviewed-by: Michael Mann <mmann78@netscape.net>