Clean up processing of encoded addresses.
Allocate the buffer at the point we fill it in, and pick the appropriate
size or let the wmem_strXXX routine do it for us.

If we aren't using an address table, just fetch the value as an IPv4
address and hand it to ip_to_str_buf() - don't fetch it in host byte
order and then *fix* it by byte-swapping (hint: on a big-endian machine,
host byte order *is* network byte order and you don't want to swap it;
not all the world's an x86).

Change-Id: I966b107271ba166ff76a5600fbc4922808e7ead1
Reviewed-on: https://code.wireshark.org/review/24159
Reviewed-by: Guy Harris <guy@alum.mit.edu>