Link with ZLIB_LIBRARIES if you link with wiretap.
Wiretap may use zlib; if it does, ZLIB_LIBRARIES is set to refer to
zlib.  On UN*X, you may be able to get away with linking a
dynamically-linked shared library with other dynamically-linked shared
libraries and not linking programs linked *with* that shared library
with those other shared libraries, but that may not work on Windows.

We link most programs that use wiretap with ZLIB_LIBRARIES; do so with
androiddump and randpktdump as well.

Bug: 14207
Change-Id: I8e94197e06f5fd0ff8c95aa509dbcc2ff2a44cd4
Reviewed-on: https://code.wireshark.org/review/24389
Reviewed-by: Guy Harris <guy@alum.mit.edu>