Handle DLT_ENC files *not* written on the same OS and byte-order host.
Some fields in the header are in the byte order of the host that wrote
them; one of them is a 32-bit AF_ value, and those are not likely ever
to be > 65535, so they should never have any of the upper 16 bits set,
and are also unlikely ever to be AF_UNSPEC, i.e. 0, so they should have
at least one of the lower 16 bits set.  This means that they will have
at least one of the upper 16 bits set iff the host that wrote the file
has the opposite byte order of the host that's reading the file; use
that to determine whether to byte-swap the address-family or flags
fields.  (The SPI field is in *network* byte order.)

Change-Id: I2d483c75d5c6bbab8fd16c5dc0a800f8710f764c
Reviewed-on: https://code.wireshark.org/review/24998
Reviewed-by: Guy Harris <guy@alum.mit.edu>